+++ tests/engine/eval_as_bool/function.rs
@@ -1,0 +1,16 @@
+use balena_template::engine::context::Context;
+use balena_template::engine::Engine;
+
+macro_rules! test_eval_err {
+    ($e:expr) => {{
+        let engine = Engine::default();
+        let context = Context::default();
+        assert!(engine.eval_as_bool(&$e.parse().unwrap(), &context).is_err());
+    }};
+}
+
+#[test]
+fn test_uuidv4() {
+    // Evaluates to String
+    test_eval_err!("uuidv4()");
+}
