+++ tests/engine/eval_as_bool/mod.rs
@@ -1,86 +1,5 @@
-use balena_template::engine::context::Context;
-use balena_template::engine::Engine;
-
+mod function;
 mod logical;
+mod math;
 mod primitive;
-
-macro_rules! test_eval_eq {
-    ($e:expr, $r:expr) => {{
-        let engine = Engine::default();
-        let context = Context::default();
-        assert_eq!(engine.eval_as_bool(&$e.parse().unwrap(), &context).unwrap(), $r);
-    }};
-}
-
-macro_rules! test_eval_err {
-    ($e:expr) => {{
-        let engine = Engine::default();
-        let context = Context::default();
-        assert!(engine.eval_as_bool(&$e.parse().unwrap(), &context).is_err());
-    }};
-}
-
-#[test]
-fn test_logical_equal() {
-    test_eval_eq!("true == true", true);
-    test_eval_eq!("1 == 1", true);
-    test_eval_eq!("2.3 == 2.3", true);
-    test_eval_eq!("`a` == `a`", true);
-    test_eval_eq!("`a` == `b`", false);
-    test_eval_eq!("`1` == 1", false);
-}
-
-#[test]
-fn test_logical_not_equal() {
-    test_eval_eq!("true != true", false);
-    test_eval_eq!("1 != 1", false);
-    test_eval_eq!("2.3 != 2.3", false);
-    test_eval_eq!("`a` != `a`", false);
-    test_eval_eq!("`a` != `b`", true);
-    test_eval_eq!("`1` != 1", true);
-}
-
-#[test]
-fn test_relational_greater_than() {
-    test_eval_eq!("1 > 2", false);
-    test_eval_eq!("3 > 2", true);
-    test_eval_eq!("3.1 > 2", true);
-}
-
-#[test]
-fn test_relational_greater_than_or_equal() {
-    test_eval_eq!("1 >= 2", false);
-    test_eval_eq!("3 >= 2", true);
-    test_eval_eq!("3.1 >= 2", true);
-    test_eval_eq!("3.1 >= 3.1", true);
-    test_eval_eq!("3 >= 3.0", true);
-}
-
-#[test]
-fn test_relational_lower_than() {
-    test_eval_eq!("1 < 2", true);
-    test_eval_eq!("3 < 2", false);
-    test_eval_eq!("3.1 < 2", false);
-    test_eval_eq!("3.1 < 3.1", false);
-    test_eval_eq!("3 < 3.0", false);
-}
-
-#[test]
-fn test_relational_lower_than_or_equal() {
-    test_eval_eq!("1 <= 2", true);
-    test_eval_eq!("3 <= 2", false);
-    test_eval_eq!("2 <= 3.1", true);
-    test_eval_eq!("3.1 <= 3.1", true);
-    test_eval_eq!("3 <= 3.0", true);
-}
-
-#[test]
-fn test_math() {
-    test_eval_err!("1 - 1");
-    test_eval_err!("2 - 1");
-}
-
-#[test]
-fn test_function() {
-    test_eval_err!("uuidv4()");
-}
+mod relational;
